+++ core/domain/src/main/java/com/blue/domain/database/todo/ChangeCheckBoxUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class ChangeCheckBoxUseCase @Inject constructor(
     private val database: TodoRepo
 ) {
-    suspend operator fun invoke(id: Int, status: Boolean) =
+    suspend operator fun invoke(id: Long, status: Boolean) =
         database.changeCheckBox(id, status)
 }
